+++ src/ray/destroy.js
@@ -7,6 +7,9 @@
  * @since 0.10.3
  */
  export function destroy() {
+    if(this.graphics)
+        this.graphics.destroy();
+
     for(let key in this) {
         delete this[key];
     }